OVERTIMEPAY _ KTH 1. 마이페이지 내 시간외 근로수당 조회 페이지 작업 중

+++ HiveWorks/src/main/java/com/dna/hiveworks/model/dto/salary/OvertimePay.java
@@ -22,7 +22,7 @@ public class OvertimePay {
 	private int emp_no;
 	private int ot_pay_per_time;
 	private int overtime;
-	private int overtimeByMonth;
+	private Double overtimeByMonth;
 	private String ot_pay_month;
 	private List<Employee> employees;
 	private String use_yn;
